training/MNISTFashionMicroservice/requirements.txt
pydantic
fastapi
pyinstaller
setuptools
uuid
uvicorn
python-multipart
pillow
python-dotenv
requests
starlette
tensorflow
matplotlib
numpy
rq
redis
